working on public link for branches 80%

+++ src/jsx/components/Services/serviceArea.js
@@ -4,8 +4,17 @@ import { Button, Modal,  Form } from "react-bootstrap";
 import axios from "axios";
 import swal from "sweetalert"
 import { useTranslation } from "react-i18next";
-
+import { useForm } from "react-hook-form";
+import { yupResolver } from '@hookform/resolvers/yup';
+import * as yup from 'yup';
 const ServiceArea = (props) => {
+    // validation
+    const schema = yup.object().shape({
+        AreaName: yup.string().required("This field is a required field"),
+    }).required();
+    const { register, handleSubmit, formState:{ errors } } = useForm({
+    resolver: yupResolver(schema)
+    });
     // for localization
     const { t } = useTranslation();
     //ID
@@ -110,22 +119,40 @@ const ServiceArea = (props) => {
     }
     // delete section 
     const deleteServiceArea= (e,id)=>{
-        e.preventDefault();
-        axios.delete(`/api/DeleteServiceAreas/${id}`).then(res=>{
-            if(res.data.status === 200){
-                swal("Success",res.data.message,"success");
-                // thisClicked.closest("tr").remove();
-            }else if(res.data.status === 404){
-                swal("Error",res.data.message,"error");
+        swal({
+            title: "Are you sure?",
+            text: "Once deleted, you will not be able to recover this imaginary file!",
+            icon: "warning",
+            buttons: [t('cancel'), t('confirm')],
+            dangerMode: true,
+          })
+          .then((willDelete) => {
+            if (willDelete) {
+                axios.delete(`/api/DeleteServiceAreas/${id}`).then(res=>{
+                    if(res.data.status === 200){
+                        swal("Success",res.data.message,"success");
+                        setServiceAreaInsert({
+                            AreaName: '',
+                            BranchID: id
+                        })
+                    }else if(res.data.status === 404){
+                        swal("Error",res.data.message,"error");
+                    }
+                   
+                });
+               
+            } else {
+              swal("Your Data is safe now!");
             }
-        });
+          });
+       
     }
     return (
       <Fragment>
          <PageTItle headingPara={t('services_area')} activeMenu={t('add_service_area')} motherMenu={t('services_area')} />
         {/* <!-- Insert  Modal --> */}
         <Modal className="fade" show={modalCentered}>
-            <Form onSubmit={saveServiceAreas} method= "POST" encType="multipart/form-data">
+            <Form onSubmit={handleSubmit(saveServiceAreas)} method= "POST" encType="multipart/form-data">
                 <Modal.Header>
                     <Modal.Title>{t('add_service_area')}</Modal.Title>
                     <Button
@@ -145,13 +172,16 @@ const ServiceArea = (props) => {
                             <label className="mb-1 "> <strong>{t('service_area')}</strong> </label>
                             <textarea
                                 type="text"
+                                {...register("AreaName")}
                                 className="form-control"
                                 placeholder={t('service_area')}
                                 name="AreaName"
-                                required
                                 onChange={handleInput}  
                                 value={serviceAreaInsert.AreaName}
                             />
+                             <div className="text-danger">
+                                 {errors.AreaName?.message}
+                              </div>
                         </div>
                 </Modal.Body>
                 <Modal.Footer>
@@ -192,7 +222,6 @@ const ServiceArea = (props) => {
                                 value={editServiceAreas.AreaName}
                             />
                         </div>
-
                 </Modal.Body>
                 <Modal.Footer>
                     <Button
